+++ Week3/Inheritance/Child.java
@@ -1,0 +1,16 @@
+public class Child extends Parent {
+    String first_name;
+
+    Child(String first_name, String lastName, String eyeColor) {
+        super(lastName, eyeColor);
+        this.first_name = first_name;
+    }
+
+    void getFullName() {
+        System.out.println("This is" + first_name + " his last name is " + last_name + ".");
+    }
+
+    void parentEyeColor() {
+        System.out.println("The parent has " + super.eye_color + " eyes.");
+    }
+}
